metaheuristics/genetics-IN_PROGRESS/requirements.txt
python-dotenv
numpy